clothingMarket/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.28
mysqlclient==2.2.4
python-dotenv==1.0.1
numpy==1.26.4 
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2